feat(pdf): generacion del PDF de la solicitud con las firmas incrustadas
Segunda mitad de la etapa 2. Con esto quedan listas las piezas que Permisos y
Vacaciones van a compartir; engancharlas al flujo es la etapa 3.

EL PDF. HtmlSolicitud arma el documento y PDFs.GenerarPdfSolicitud lo
convierte. Dos cosas del marcado se ven anticuadas a proposito:

  - Maquetacion con tablas y estilos en linea. Pechkin 0.5.8 envuelve una
    version vieja de wkhtmltopdf que no entiende flexbox ni grid, y cuando no
    entiende no da error: saca la pagina desarmada.
  - Las imagenes van por ruta de archivo con file:///, no como data URI. Es el
    mismo camino que ya usaba GenerarCodigoQR, que devuelve una ruta
    justamente porque esa version no resuelve base64 embebido de forma
    confiable. Los trazos se escriben a una carpeta temporal, se convierten y
    se borran; el original vive en la base.

METODO NUEVO Y NO REUSAR EL VIEJO. EnvioCorreoEncuesta hace casi lo mismo pero
no servia: su variable de resultado se declara en false y no se toca nunca, asi
que quien la llama no puede distinguir un PDF creado de uno que no; arma el
nombre con "hh:mm:ss", hora de 12 sin AM/PM, de modo que dos documentos de la
01:00 y las 13:00 del mismo dia se pisan; y se llama EnvioCorreoEncuesta sin
enviar correos ni tener que ver con encuestas. No se la toco porque hay tres
pantallas colgando de ella.

FIRMA. La accion GuardarFirmaSolicitud recibe el trazo como data URI y lo
convierte a bytes. Nombre, cargo y cedula NO llegan de la pantalla: los copia
el procedimiento desde R_Usuarios con el codigo de la sesion, para que no se
puedan digitar. Se guardan IP y agente; la IP sale de X-Forwarded-For cuando
hay proxy, porque UserHostAddress daria la del balanceador.

IDVACACIONES ES BIGINT. Lo confirme contra la base: la columna es bigint y yo
habia declarado la tabla de firmas en int. El maximo actual es 22837 y cabria,
pero una columna de union con tipo distinto al origen provoca conversiones
implicitas que arruinan los indices cuando crece. Como la tabla todavia no
existia, alinearla salio gratis. El cargador de solicitudes sigue recibiendo
int, asi que ahi va un Convert que lanza si no cabe: mejor un error visible
que un id truncado en silencio.

Verificado contra la base, esta vez si:
  - El script corrio completo. Cargo quedo varchar(128), IdVacaciones bigint,
    la restriccion unica esta, y los dos SP de usuarios traen y auditan Cargo.
  - Las cuatro validaciones rechazan: rol invalido, decision invalida,
    solicitud inexistente y usuario inexistente.
  - Una firma real con trazo: se registra, el comentario se recorta, el rol en
    minusculas se normaliza, el doble clic responde "ya estaba firmado" sin
    duplicar, y el base64 vuelve identico byte por byte. Todo dentro de una
    transaccion revertida: la tabla quedo en cero.

NO verificado: la conversion de Pechkin en si. Necesita HttpContext, o sea el
sitio andando. Compila y las dependencias de datos estan probadas, pero que
wkhtmltopdf resuelva los file:/// y saque la pagina armada hay que verlo en el
navegador.

+++ CapaDato/DaoFirmaSolicitud.cs
@@ -25,7 +25,7 @@ namespace CapaDato
             using (SqlCommand cmd = new SqlCommand("Sp_RTA_GuardarFirmaSolicitud", cnx))
             {
                 cmd.CommandType = CommandType.StoredProcedure;
-                cmd.Parameters.Add("@IdVacaciones", SqlDbType.Int).Value = firma.IdVacaciones;
+                cmd.Parameters.Add("@IdVacaciones", SqlDbType.BigInt).Value = firma.IdVacaciones;
                 cmd.Parameters.Add("@Rol", SqlDbType.VarChar, 20).Value = firma.Rol ?? string.Empty;
                 cmd.Parameters.Add("@Secuencia", SqlDbType.Int).Value = firma.Secuencia <= 0 ? 1 : firma.Secuencia;
                 cmd.Parameters.Add("@Decision", SqlDbType.VarChar, 10).Value = firma.Decision ?? string.Empty;
@@ -55,7 +55,7 @@ namespace CapaDato
         }
 
         /// <summary>Firmas de una solicitud, en el orden del flujo.</summary>
-        public static List<EntFirmaSolicitud> Listar(int idVacaciones)
+        public static List<EntFirmaSolicitud> Listar(long idVacaciones)
         {
             List<EntFirmaSolicitud> lista = new List<EntFirmaSolicitud>();
             DaoReporTareaAranda conexion = new DaoReporTareaAranda();
@@ -64,7 +64,7 @@ namespace CapaDato
             using (SqlCommand cmd = new SqlCommand("Sp_RTA_ListarFirmasSolicitud", cnx))
             {
                 cmd.CommandType = CommandType.StoredProcedure;
-                cmd.Parameters.Add("@IdVacaciones", SqlDbType.Int).Value = idVacaciones;
+                cmd.Parameters.Add("@IdVacaciones", SqlDbType.BigInt).Value = idVacaciones;
                 cnx.Open();
 
                 using (SqlDataReader dr = cmd.ExecuteReader())
@@ -74,7 +74,7 @@ namespace CapaDato
                         lista.Add(new EntFirmaSolicitud()
                         {
                             IdFirma = Convert.ToInt32(dr["IdFirma"]),
-                            IdVacaciones = Convert.ToInt32(dr["IdVacaciones"]),
+                            IdVacaciones = Convert.ToInt64(dr["IdVacaciones"]),
                             Rol = dr["Rol"].ToString(),
                             Secuencia = Convert.ToInt32(dr["Secuencia"]),
                             Decision = dr["Decision"].ToString(),
